+++ CHANGES.md
@@ -2,5 +2,7 @@
 
 ## 0.8.1-pre
 
+- Same version of SQLitePlugin.js working in both main thread and worker thread(s)
+- Fix iOS version to work in main thread using Cordova again
 - Fix conversion warnings in iOS version
 - Fix readTransaction to skip BEGIN/COMMIT/ROLLBACK